+++ intervals/interval.py
@@ -50,11 +50,6 @@ except ImportError:
 from .exc import IllegalArgument, IntervalException, RangeBoundsException
 from .parser import IntervalParser, IntervalStringParser
 
-try:
-    string_types = basestring,  # Python 2
-except NameError:
-    string_types = str,  # Python 3
-
 
 def is_number(number):
     return isinstance(number, (float, int, Decimal))
@@ -133,7 +128,7 @@ def coerce_interval(func):
                 return NotImplemented
         try:
             arg = type(self)(self.type(arg))
-        except (ValueError, TypeError):
+        except (ValueError, TypeError, OverflowError):
             pass
         return func(self, arg)
     return wrapper
@@ -204,7 +199,7 @@ class AbstractInterval(object):
             30
 
         """
-        if isinstance(bounds, string_types):
+        if isinstance(bounds, str):
             raise TypeError(
                 'First argument should be a list or tuple. If you wish to '
                 'initialize an interval from string, use from_string factory '
@@ -334,7 +329,7 @@ class AbstractInterval(object):
             return value
         elif isinstance(value, self.type):
             return value
-        elif isinstance(value, string_types):
+        elif isinstance(value, str):
             return self.coerce_string(value)
         else:
             return self.coerce_obj(value)
@@ -489,6 +484,11 @@ class AbstractInterval(object):
 
     @property
     def length(self):
+        if self.discrete:
+            if not self:
+                return 0
+            if not self.lower_inc or not self.upper_inc:
+                return canonicalize(self, lower_inc=True, upper_inc=True).length
         return abs(self.upper - self.lower)
 
     @property